solved Conflicts:
	backend/de.metas.business/src/main/java/de/metas/product/IProductDAO.java
	backend/de.metas.business/src/main/java/de/metas/product/impl/ProductDAO.java

+++ backend/de.metas.adempiere.adempiere/base/src/main/java-gen/org/compiere/model/I_M_Product.java
@@ -1863,21 +1863,21 @@ public interface I_M_Product
 	 * Set Regaltiefe.
 	 * Shelf depth required
 	 *
-	 * <br>Type: Integer
+	 * <br>Type: Amount
 	 * <br>Mandatory: false
 	 * <br>Virtual Column: false
 	 */
-	void setShelfDepth (int ShelfDepth);
+	void setShelfDepth (BigDecimal ShelfDepth);
 
 	/**
 	 * Get Regaltiefe.
 	 * Shelf depth required
 	 *
-	 * <br>Type: Integer
+	 * <br>Type: Amount
 	 * <br>Mandatory: false
 	 * <br>Virtual Column: false
 	 */
-	int getShelfDepth();
+	BigDecimal getShelfDepth();
 
 	ModelColumn<I_M_Product, Object> COLUMN_ShelfDepth = new ModelColumn<>(I_M_Product.class, "ShelfDepth", null);
 	String COLUMNNAME_ShelfDepth = "ShelfDepth";
@@ -1909,21 +1909,21 @@ public interface I_M_Product
 	 * Set Regalbreite.
 	 * Shelf width required
 	 *
-	 * <br>Type: Integer
+	 * <br>Type: Amount
 	 * <br>Mandatory: false
 	 * <br>Virtual Column: false
 	 */
-	void setShelfWidth (int ShelfWidth);
+	void setShelfWidth (BigDecimal ShelfWidth);
 
 	/**
 	 * Get Regalbreite.
 	 * Shelf width required
 	 *
-	 * <br>Type: Integer
+	 * <br>Type: Amount
 	 * <br>Mandatory: false
 	 * <br>Virtual Column: false
 	 */
-	int getShelfWidth();
+	BigDecimal getShelfWidth();
 
 	ModelColumn<I_M_Product, Object> COLUMN_ShelfWidth = new ModelColumn<>(I_M_Product.class, "ShelfWidth", null);
 	String COLUMNNAME_ShelfWidth = "ShelfWidth";
